refactor and merge spring apps
